+++ clone/config/auth.php
@@ -6,8 +6,36 @@
  *             See https://github.com/faithfm/laravel-auth0-pattern for more details.
  */
 
- // Note: auth0/login v7's implementation is now very reliant on the specific naming of guard + driver + provider = 'auth0'.  
- //    (This can be a little confusing to understand/debug since each performs a different function, yet all three are identically named.)
+
+ /** 
+ Note: auth0/login v7's implementation is now very reliant on the specific naming of guard + driver + provider = 'auth0'.
+     (This can be a little confusing to understand/debug since each performs a different function, yet all three are identically named.)
+     
+    This later led to other issues that have been temporarily 'fixed' by creating duplicate 'web' + 'auth0' guards to keep both Laravel and Auth0 happy  (see below)
+
+    EXPLANATION
+    The specific config('auth.guards.auth0.provider') that has been coded into 'auth0/login' src/Auth/Guard.php prevents us from naming our default guard 'web' 
+    using an 'auth0' driver... (our preferred approach in laravel-auth0-pattern v1.x).
+
+    ...but if we try to use 'auth0.authenticate' or 'auth0.authenticate.default' guards...
+    (as per suggestions in the QuickStart: https://auth0.com/docs/quickstart/webapp/laravel#protecting-routes)
+
+    ...then we experience side-effects/errors when the default 'web' guard is not used in the middleware.
+    (see others experiencing same issue here: https://community.auth0.com/t/implementing-auth0-in-addition-to-existing-laravel-auth/91225)
+
+    One solution would be to fork our own version of the 'auth0/login' repo that looks up the actual provider (rather than assuming specific hard-coded naming), however 'auth0/login' is currently unstable and somewhat difficult to debug, and we found it easier to "cheat" by creating duplicate 'web' + 'auth0' guards in config/auth.php - which:
+    1. Keeps Laravel happy by allowing us to use the default 'web' guards in our middleware
+    2. Keeps 'auth0/login' hard-coded settings happy because the duplicate 'auth0' guard exists in config/auth.php.
+
+    Note: 
+    The actual error is caused when we try to use the 'auth0.authenticate' guard instead of the 'web' guard for our API routes middleware (in RouteServiceProvider.php).  
+
+    It seemed to work ok for actual web routes, but didn't work properly as an additional guard for API routes.  We weren't able to get to the bottom of the issue, but an exception is raised when a null user is returned in some circumstances because it seems to not be able to retrieve a stateful user - see code sections:
+    - src/Auth/Guard.php functions: user(), getUserFromSession()
+    - src/Http/Middleware/Stateful/AuthenticateOptional.php functions: handle()
+
+  */
+
 
 return [
 
@@ -46,6 +74,12 @@ return [
 
     'guards' => [
         'auth0' => [
+            'driver' => 'auth0',
+            'provider' => 'auth0',
+        ],
+
+        // duplicate of 'auth0' guard required as temporary bug-fix - see notes at top of file:
+        'web' => [
             'driver' => 'auth0',
             'provider' => 'auth0',
         ],